+++ assets/javascripts/kitten/karl/pages/homepage-2020/engagements-carousel/stories.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { EngagementsCarousel } from './index'
+
+export default {
+  component: EngagementsCarousel,
+  title: 'pages/Homepage 2020/EngagementsCarousel',
+  parameters: {
+    component: EngagementsCarousel,
+  },
+}
+
+export const DefaultProps = () => <EngagementsCarousel />