+++ assets/javascripts/kitten/karl/pages/homepage-2020/news/stories.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import HomepageNews from './index'
+
+export default {
+  component: HomepageNews,
+  title: 'pages/Homepage 2020/HomepageNews',
+  parameters: {
+    component: HomepageNews,
+  },
+}
+
+export const DefaultProps = () => <HomepageNews />