+++ assets/javascripts/kitten/karl/pages/homepage-2020/popular-projects/stories.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { PopularProjects } from './index'
+
+export default {
+  component: PopularProjects,
+  title: 'pages/Homepage 2020/PopularProjects',
+  parameters: {
+    component: PopularProjects,
+  },
+}
+
+export const DefaultProps = () => <PopularProjects />