+++ src/components/graphics/illustrations/computer-illustration.js
@@ -1,0 +1,148 @@
+"use strict";
+
+var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefault");
+
+exports.__esModule = true;
+exports.ComputerIllustration = void 0;
+
+var _extends2 = _interopRequireDefault(require("@babel/runtime/helpers/extends"));
+
+var _react = _interopRequireDefault(require("react"));
+
+var _propTypes = _interopRequireDefault(require("prop-types"));
+
+const ComputerIllustration = _ref => {
+  let {
+    title,
+    ...props
+  } = _ref;
+  return /*#__PURE__*/_react.default.createElement("svg", (0, _extends2.default)({
+    xmlns: "http://www.w3.org/2000/svg",
+    width: "1289",
+    height: "1123",
+    fill: "none",
+    viewBox: "0 0 1289 1123"
+  }, props), title && /*#__PURE__*/_react.default.createElement("title", null, title), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#0091FF",
+    d: "M3.234 326.584V88.389C3.234 41.6 41.156 3.675 87.94 3.675H1163.6v866.622H87.793c-46.696 0-84.544-37.852-84.544-84.553v-459.16h-.015Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#7BB3FF",
+    d: "M85.103 3.675H1161.35v115.069H5.498V83.289c0-43.938 35.672-79.614 79.605-79.614Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#D2D9D9",
+    d: "M2.602 734.163H1165.12v136.134H89.233c-47.814 0-86.631-38.822-86.631-86.64v-49.494Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M3.674 326.584V88.389c0-46.789 37.921-84.714 84.705-84.714H1164.04v866.622H88.218c-46.696 0-84.544-37.852-84.544-84.553v-459.16Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#000",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M1169.02 3.675h55.5c33.57 0 60.82 27.253 60.82 60.827V809.47c0 33.574-27.25 60.827-60.82 60.827h-55.5V3.675Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#D2D9D9",
+    d: "M996.316 1119.33H170.514v-18.19c0-22.96 18.608-41.58 41.566-41.58h777.357"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M996.316 1119.33H170.514v-18.19c0-22.96 18.608-41.58 41.566-41.58h777.357"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#000",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M1052.93 1101.14v18.19h-55.718v-18.19c0-22.96-18.623-41.58-41.581-41.58h55.719c22.96 0 41.58 18.62 41.58 41.58Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#D2D9D9",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M705.042 870.297H462.684v189.263h242.358V870.297Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M6.496 734.163H1169.02M602.89 840.427c20.172 0 36.525-16.354 36.525-36.529 0-20.174-16.353-36.528-36.525-36.528-20.172 0-36.525 16.354-36.525 36.528 0 20.175 16.353 36.529 36.525 36.529Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#fff",
+    d: "M1026.61 196.594h-906.1c-23.362 0-40.301 18.941-40.301 42.306v57.946c0 23.365 16.939 42.306 40.301 42.306h906.1c23.36 0 42.3-18.941 42.3-42.306V238.9c0-23.365-18.94-42.306-42.3-42.306Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M5.498 118.744H1161.35"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#F60044",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M103.578 83.171c11.738 0 21.253-9.517 21.253-21.256s-9.515-21.256-21.253-21.256-21.254 9.517-21.254 21.256c0 11.74 9.516 21.256 21.254 21.256Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#FFB500",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M176.363 83.171c11.738 0 21.253-9.517 21.253-21.256s-9.515-21.256-21.253-21.256-21.254 9.517-21.254 21.256c0 11.74 9.516 21.256 21.254 21.256Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#08A777",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M249.148 83.171c11.738 0 21.254-9.517 21.254-21.256s-9.516-21.256-21.254-21.256c-11.738 0-21.253 9.517-21.253 21.256 0 11.74 9.515 21.256 21.253 21.256Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#fff",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M1062.93 42.13H358.429c-11.738 0-21.253 9.516-21.253 21.255 0 11.74 9.515 21.256 21.253 21.256h704.501c11.74 0 21.25-9.517 21.25-21.256 0-11.74-9.51-21.256-21.25-21.256Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#08A777",
+    d: "M120.907 196.015H244V338H120.907C99.442 338 82 320.944 82 299.92v-65.84C82 213.071 99.427 196 120.907 196v.015Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M1030.02 195.639h-906.1c-23.362 0-42.3 18.941-42.3 42.305v57.947c0 23.365 18.938 42.305 42.3 42.305h906.1c23.36 0 42.3-18.94 42.3-42.305v-57.947c0-23.364-18.94-42.305-42.3-42.305Zm-785.825.955v142.558"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#fff",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M158.05 288.379c16.722 0 30.278-13.557 30.278-30.281 0-16.724-13.556-30.282-30.278-30.282s-30.279 13.558-30.279 30.282 13.557 30.281 30.279 30.281Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "m179.303 279.662 25.236 25.24"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#6EB2FF",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M583.414 390.88h-498.4v292.76h498.4V390.88Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#FFB8CB",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "M1068.91 390.88H626.994v123.816h441.916V390.88Zm0 166.798H626.994V683.64h441.916V557.678Z"
+  }), /*#__PURE__*/_react.default.createElement("path", {
+    fill: "#F60044",
+    stroke: "#000",
+    strokeMiterlimit: "10",
+    strokeWidth: "5",
+    d: "m470.856 531.792-35.349 18.801a2.097 2.097 0 0 0-.499 3.322l50.473 50.479a2.101 2.101 0 0 1 0 2.954l-17.961 17.963a2.101 2.101 0 0 1-2.954 0l-50.474-50.478c-.985-.985-2.66-.735-3.322.499l-18.799 35.353c-.867 1.632-3.248 1.411-3.821-.338l-38.392-118.142c-.529-1.617 1.014-3.16 2.631-2.631l118.129 38.396c1.749.573 1.955 2.955.338 3.822Z"
+  }));
+};
+
+exports.ComputerIllustration = ComputerIllustration;
+ComputerIllustration.propTypes = {
+  title: _propTypes.default.string
+};
+ComputerIllustration.defaultProps = {
+  title: null
+};